pyproject.toml
[tool.poetry]
name = "image-to-table"
version = "0.1.0"
description = ""
authors = ["Vegard Stikbakke <vegard.stikbakke@gmail.com>"]

[tool.black]
line-length = 120
target_version = ['py37']
include = '\.py$'

[tool.isort]
line_length=120                # corresponds to -w  flag
multi_line_output=3            # corresponds to -m  flag
include_trailing_comma=true    # corresponds to -tc flag
skip_glob = '^((?!py$).)*$'    # isort all Python files
known_third_party = ["cv2", "flask", "google", "numpy", "opencv_wrapper", "pytest", "scipy", "werkzeug"]

[tool.poetry.dependencies]
python = "^3.7"
opencv-python = "^4.0"
google-cloud-vision = "^0.39.0"
matplotlib = "^3.1"
grpcio = "^1.24"
numpy = "^1.16"
shapely = "^1.6"
csvprint = "^0.6.1"
opencv-wrapper = "^0.2.3"
scipy = "^1.3"
flask = "^1.1"
pytest = "^5.3"

[tool.poetry.dev-dependencies]
ipython = "^7.8"
isort = "^4.3"
seed-isort-config = "^1.9"
black = {version = "^18.3-alpha.0", allows-prereleases = true}
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
